Update playground
Divide example into three separate pages and add more comments
